validate getters/setters for key variables

+++ contracts/rubiconPools/BathPair.sol
@@ -122,25 +122,22 @@ contract BathPair {
 
     // TODO: enforce oracle sanity check to avoid order book manipulation before permissionless strategists
     function getMidpointPrice() internal returns (int128) {
-        uint256 bestAskID =
-            RubiconMarket(RubiconMarketAddress).getBestOffer(
-                ERC20(underlyingAsset),
-                ERC20(underlyingQuote)
-            );
-        uint256 bestBidID =
-            RubiconMarket(RubiconMarketAddress).getBestOffer(
-                ERC20(underlyingQuote),
-                ERC20(underlyingAsset)
-            );
+        uint256 bestAskID = RubiconMarket(RubiconMarketAddress).getBestOffer(
+            ERC20(underlyingAsset),
+            ERC20(underlyingQuote)
+        );
+        uint256 bestBidID = RubiconMarket(RubiconMarketAddress).getBestOffer(
+            ERC20(underlyingQuote),
+            ERC20(underlyingAsset)
+        );
 
         order memory bestAsk = getOfferInfo(bestAskID);
         order memory bestBid = getOfferInfo(bestBidID);
-        int128 midpoint =
-            ABDKMath64x64.divu(
-                ((bestAsk.buy_amt / bestAsk.pay_amt) +
-                    (bestBid.pay_amt / bestBid.buy_amt)),
-                2
-            );
+        int128 midpoint = ABDKMath64x64.divu(
+            ((bestAsk.buy_amt / bestAsk.pay_amt) +
+                (bestBid.pay_amt / bestBid.buy_amt)),
+            2
+        );
         // ((bestAsk.buy_amt / bestAsk.pay_amt) +
         //     (bestBid.pay_amt / bestBid.buy_amt)) / 2;
         emit LogNoteI("midpoint calculated:", midpoint);
@@ -160,16 +157,14 @@ contract BathPair {
             (askN * bidD) < (bidN * askD),
             "there is not a spread on strategist's pair trade"
         );
-        uint256 bestAskID =
-            RubiconMarket(RubiconMarketAddress).getBestOffer(
-                ERC20(underlyingAsset),
-                ERC20(underlyingQuote)
-            );
-        uint256 bestBidID =
-            RubiconMarket(RubiconMarketAddress).getBestOffer(
-                ERC20(underlyingQuote),
-                ERC20(underlyingAsset)
-            );
+        uint256 bestAskID = RubiconMarket(RubiconMarketAddress).getBestOffer(
+            ERC20(underlyingAsset),
+            ERC20(underlyingQuote)
+        );
+        uint256 bestBidID = RubiconMarket(RubiconMarketAddress).getBestOffer(
+            ERC20(underlyingQuote),
+            ERC20(underlyingAsset)
+        );
 
         order memory bestAsk = getOfferInfo(bestAskID);
         order memory bestBid = getOfferInfo(bestBidID);
@@ -199,10 +194,12 @@ contract BathPair {
 
     // Returns filled liquidity to the correct bath pool
     function rebalancePair() internal {
-        uint256 bathAssetYield =
-            ERC20(underlyingQuote).balanceOf(bathAssetAddress);
-        uint256 bathQuoteYield =
-            ERC20(underlyingAsset).balanceOf(bathQuoteAddress);
+        uint256 bathAssetYield = ERC20(underlyingQuote).balanceOf(
+            bathAssetAddress
+        );
+        uint256 bathQuoteYield = ERC20(underlyingAsset).balanceOf(
+            bathQuoteAddress
+        );
         uint8 stratReward = BathHouse(bathHouse).getPropToStrats(address(this));
         if (bathAssetYield > 0) {
             BathToken(bathAssetAddress).rebalance(
@@ -226,16 +223,16 @@ contract BathPair {
         uint256 fillCountA = strategist2FillsAsset[msg.sender];
         uint256 fillCountQ = strategist2FillsQuote[msg.sender];
         if (fillCountA > 0) {
-            uint256 booty =
-                (fillCountA * ERC20(underlyingAsset).balanceOf(address(this))) /
-                    totalAssetFills;
+            uint256 booty = (fillCountA *
+                ERC20(underlyingAsset).balanceOf(address(this))) /
+                totalAssetFills;
             IERC20(underlyingAsset).transfer(msg.sender, booty);
             totalAssetFills -= fillCountA;
         }
         if (fillCountQ > 0) {
-            uint256 booty =
-                (fillCountQ * ERC20(underlyingQuote).balanceOf(address(this))) /
-                    totalQuoteFills;
+            uint256 booty = (fillCountQ *
+                ERC20(underlyingQuote).balanceOf(address(this))) /
+                totalQuoteFills;
             IERC20(underlyingQuote).transfer(msg.sender, booty);
             totalQuoteFills -= fillCountQ;
         }
@@ -266,8 +263,10 @@ contract BathPair {
         }
     }
 
-    function removeElement(uint index) internal {
-        outstandingPairIDs[index] = outstandingPairIDs[outstandingPairIDs.length - 1];
+    function removeElement(uint256 index) internal {
+        outstandingPairIDs[index] = outstandingPairIDs[
+            outstandingPairIDs.length - 1
+        ];
         outstandingPairIDs.pop();
     }
 
@@ -342,13 +341,17 @@ contract BathPair {
 
     // Get offer info from Rubicon Market
     function getOfferInfo(uint256 id) internal view returns (order memory) {
-        (uint256 ask_amt, ERC20 ask_gem, uint256 bid_amt, ERC20 bid_gem) =
-            RubiconMarket(RubiconMarketAddress).getOffer(id);
+        (
+            uint256 ask_amt,
+            ERC20 ask_gem,
+            uint256 bid_amt,
+            ERC20 bid_gem
+        ) = RubiconMarket(RubiconMarketAddress).getOffer(id);
         order memory offerInfo = order(ask_amt, ask_gem, bid_amt, bid_gem);
         return offerInfo;
     }
 
-    function getOutstandingPairCount() external view returns(uint){
+    function getOutstandingPairCount() external view returns (uint256) {
         return outstandingPairIDs.length;
     }
 
@@ -360,7 +363,10 @@ contract BathPair {
         require(asset == underlyingAsset || asset == underlyingQuote);
         uint256 maxOrderSizeProportion = 50; //in percentage points of underlying
         uint256 underlyingBalance = IERC20(asset).balanceOf(bathTokenAddress);
-        require(underlyingBalance > 0, "no bathToken liquidity to calculate max orderSize permissable");
+        require(
+            underlyingBalance > 0,
+            "no bathToken liquidity to calculate max orderSize permissable"
+        );
         // emit LogNote("underlyingBalance", underlyingBalance);
         // emit LogNote(
         //     "underlying other",
@@ -373,11 +379,10 @@ contract BathPair {
         // if the asset/quote is overweighted: underlyingBalance / (Proportion of quote allocated to pair) * underlyingQuote balance
         if (asset == underlyingAsset) {
             // uint ratio = underlyingBalance / IERC20(underlyingQuote).balanceOf(bathQuoteAddress); //this ratio should equal price
-            int128 ratio =
-                ABDKMath64x64.divu(
-                    underlyingBalance,
-                    IERC20(underlyingQuote).balanceOf(bathQuoteAddress)
-                );
+            int128 ratio = ABDKMath64x64.divu(
+                underlyingBalance,
+                IERC20(underlyingQuote).balanceOf(bathQuoteAddress)
+            );
             // emit LogNoteI("ratio", ratio); // this number divided by 2**64 is correct!
             if (ABDKMath64x64.mul(ratio, getMidpointPrice()) > (2**64)) {
                 // bid at maxSize
@@ -388,24 +393,24 @@ contract BathPair {
                 return (maxOrderSizeProportion * underlyingBalance) / 100;
             } else {
                 // return dynamic order size
-                uint256 maxSize =
-                    (maxOrderSizeProportion * underlyingBalance) / 100; // Correct!
+                uint256 maxSize = (maxOrderSizeProportion * underlyingBalance) /
+                    100; // Correct!
                 // emit LogNote("raw maxSize", maxSize);
                 // int128 e = ABDKMath64x64.divu(SafeMathE.eN(), SafeMathE.eD()); //Correct as a int128!
                 // emit LogNoteI("e", e);
-                int128 shapeFactor =
-                    ABDKMath64x64.exp(ABDKMath64x64.mul(shapeCoef, ratio));
+                int128 shapeFactor = ABDKMath64x64.exp(
+                    ABDKMath64x64.mul(shapeCoef, ratio)
+                );
                 // emit LogNoteI("raised to the", shapeFactor);
                 uint256 dynamicSize = ABDKMath64x64.mulu(shapeFactor, maxSize);
                 // emit LogNote("dynamic maxSize Asset", dynamicSize); //
                 return dynamicSize;
             }
         } else if (asset == underlyingQuote) {
-            int128 ratio =
-                ABDKMath64x64.divu(
-                    underlyingBalance,
-                    IERC20(underlyingAsset).balanceOf(bathAssetAddress)
-                );
+            int128 ratio = ABDKMath64x64.divu(
+                underlyingBalance,
+                IERC20(underlyingAsset).balanceOf(bathAssetAddress)
+            );
             if (ABDKMath64x64.div(ratio, getMidpointPrice()) > (2**64)) {
                 // bid at maxSize
                 // emit LogNote(
@@ -415,13 +420,14 @@ contract BathPair {
                 return (maxOrderSizeProportion * underlyingBalance) / 100;
             } else {
                 // return dynamic order size
-                uint256 maxSize =
-                    (maxOrderSizeProportion * underlyingBalance) / 100; // Correct! 48000000000000000000
+                uint256 maxSize = (maxOrderSizeProportion * underlyingBalance) /
+                    100; // Correct! 48000000000000000000
                 // emit LogNote("raw maxSize", maxSize);
                 // int128 e = ABDKMath64x64.divu(SafeMathE.eN(), SafeMathE.eD()); //Correct as a int128!
                 // emit LogNoteI("e", e);
-                int128 shapeFactor =
-                    ABDKMath64x64.exp(ABDKMath64x64.mul(shapeCoef, ratio));
+                int128 shapeFactor = ABDKMath64x64.exp(
+                    ABDKMath64x64.mul(shapeCoef, ratio)
+                );
                 // emit LogNoteI("raised to the", shapeFactor);
                 uint256 dynamicSize = ABDKMath64x64.mulu(shapeFactor, maxSize);
                 // emit LogNote("dynamic maxSize Asset", dynamicSize); // 45728245133630216043
@@ -444,7 +450,10 @@ contract BathPair {
         internal
         returns (uint256[3] memory)
     {
-        require(outstandingPairIDs[outstandingPairIDs.length - 1][2] == block.timestamp);
+        require(
+            outstandingPairIDs[outstandingPairIDs.length - 1][2] ==
+                block.timestamp
+        );
         IDs2strategist[
             outstandingPairIDs[outstandingPairIDs.length - 1][0]
         ] = strategist;
@@ -454,7 +463,7 @@ contract BathPair {
         return outstandingPairIDs[outstandingPairIDs.length - 1];
     }
 
-    function getLastTradeIDs() external view returns(uint[3] memory) {
+    function getLastTradeIDs() external view returns (uint256[3] memory) {
         return outstandingPairIDs[outstandingPairIDs.length - 1];
     }
